requirements.txt
openai>=1.45.0
Pillow>=10.4.0
numpy>=2.1.0
pymupdf>=1.24.9
pandas>=2.2.2
tqdm>=4.66.5
python-dotenv>=1.0.1
jsonschema>=4.23.0
pytest>=8.3.3
